+++ deepxube/base/pathfind_fns.py
@@ -11,7 +11,7 @@ from deepxube.base.domain import State, Action, Goal, Domain
 from deepxube.base.nnet import DeepXubeNNet, HeurNNet, PolicyNNet
 from deepxube.base.nnet_input import NNetInput, StateGoalIn, PolicyNNetIn
 from deepxube.factories.nnet_factory import deepxube_nnet_factory
-from deepxube.factories.nnet_input_factory import get_nnet_input_t
+from deepxube.factories.nnet_input_factory import get_nnet_input_t, get_nnet_input_from_arg
 from deepxube.pytorch.nnet_utils import NNetPar, NNF_T, PROCESSED_T, ProcessedInput
 
 
@@ -112,12 +112,12 @@ class DeepXubeNNetPar(NNetPar[NNF_T, PROCESSED_T], Generic[NNF_T, PROCESSED_T, D
 
         return None
 
-    def __init__(self, domain: D, domain_name: str, nnet_input_name: Optional[str], nnet_name_args: Optional[str], **kwargs: Any):
+    def __init__(self, domain: D, domain_name: str, nnet_input_name_args: Optional[str], nnet_name_args: Optional[str], **kwargs: Any):
 
         nnet_input_t: Optional[Type[NNetInput]] = None
         nnet_t: Optional[Type[DeepXubeNNet]] = None
-        if nnet_input_name is not None:
-            nnet_input_t = get_nnet_input_t(domain_name, nnet_input_name)
+        if nnet_input_name_args is not None:
+            nnet_input_t = get_nnet_input_t(domain_name, get_name_args(nnet_input_name_args)[0])
         if nnet_name_args is not None:
             nnet_t = deepxube_nnet_factory.get_type(get_name_args(nnet_name_args)[0])
 
@@ -127,7 +127,7 @@ class DeepXubeNNetPar(NNetPar[NNF_T, PROCESSED_T], Generic[NNF_T, PROCESSED_T, D
 
         self.domain: D = domain
         self.domain_name: str = domain_name
-        self.nnet_input_name: Optional[str] = nnet_input_name
+        self.nnet_input_name_args: Optional[str] = nnet_input_name_args
         self.nnet_name_args: Optional[str] = nnet_name_args
 
         super().__init__(**kwargs)
@@ -155,8 +155,8 @@ class DeepXubeNNetPar(NNetPar[NNF_T, PROCESSED_T], Generic[NNF_T, PROCESSED_T, D
 
     def _get_nnet_input(self) -> NNInP:
         if self.nnet_input is None:
-            assert self.nnet_input_name is not None
-            self.nnet_input = cast(NNInP, get_nnet_input_t(self.domain_name, self.nnet_input_name)(domain=self.domain))
+            assert self.nnet_input_name_args is not None
+            self.nnet_input = cast(NNInP, get_nnet_input_from_arg(self.domain, self.domain_name, self.nnet_input_name_args))
 
         assert self.nnet_input is not None
         return self.nnet_input
@@ -166,11 +166,11 @@ class DeepXubeNNetPar(NNetPar[NNF_T, PROCESSED_T], Generic[NNF_T, PROCESSED_T, D
         return self.__dict__
 
     def __repr__(self) -> str:
-        if self.nnet_input_name is None:
+        if self.nnet_input_name_args is None:
             return f"{type(self).__name__}()"
         else:
-            assert self.nnet_name_args is not None
-            return f"NNetInputType: {get_nnet_input_t(self.domain_name, self.nnet_input_name)} (name: {self.nnet_input_name})\n{super().__repr__()}"
+            nnet_input_name: str = get_name_args(self.nnet_input_name_args)[0]
+            return f"NNetInputType: {get_nnet_input_t(self.domain_name, nnet_input_name)} (name: {nnet_input_name})\n{super().__repr__()}"
 
 
 H = TypeVar('H', bound=HeurFn)