+++ deepxube/factories/nnet_input_factory.py
@@ -1,5 +1,6 @@
-from typing import Dict, Type, Callable, List
+from typing import Dict, Type, Callable, List, Any
 
+from deepxube.utils.command_line_utils import get_name_args
 from deepxube.base.nnet_input import NNetInput, DynamicNNetInput
 from deepxube.base.domain import Domain
 from deepxube.base.factory import Factory, Parser
@@ -43,3 +44,11 @@ def register_nnet_input_dynamic() -> None:
             nnet_input_t_dict: Dict[str, Type[NNetInput]] = domain_t.get_dynamic_nnet_inputs()
             for nnet_input_name, nnet_input_t in nnet_input_t_dict.items():
                 register_nnet_input(domain_name, f"{nnet_input_name}")(nnet_input_t)
+
+
+def get_nnet_input_from_arg(domain: Domain, domain_name: str, nnet_input_name_args: str) -> NNetInput:
+    nnet_input_name, nnet_input_args = get_name_args(nnet_input_name_args)
+    nnet_input_factory: Factory[NNetInput] = _nnet_input_registry[domain_name]
+    nnet_input_kwargs: Dict[str, Any] = nnet_input_factory.get_kwargs(nnet_input_name, nnet_input_args)
+    nnet_input_kwargs["domain"] = domain
+    return nnet_input_factory.build_class(nnet_input_name, nnet_input_kwargs)